Make weight ckpt part of full ckpt
