+++ src/prime_rl/trainer/sft/config.py
@@ -10,7 +10,6 @@ from prime_rl.trainer.config import (
     ModelConfig,
     OptimizerConfigType,
     SchedulerConfigType,
-    WeightCheckpointConfig,
 )
 from prime_rl.utils.config import LogConfig, WandbMonitorConfig
 from prime_rl.utils.pydantic_config import BaseConfig, BaseSettings
@@ -112,9 +111,6 @@ class SFTTrainerConfig(BaseSettings):
     # The checkpoint configuration
     ckpt: CheckpointConfig | None = None
 
-    # The weight checkpoint configuration
-    weights: WeightCheckpointConfig | None = None
-
     # The logging configuration
     log: LogConfig = LogConfig()
 
@@ -192,28 +188,8 @@ class SFTTrainerConfig(BaseSettings):
         return self
 
     @model_validator(mode="after")
-    def validate_ckpt_managers(self):
-        # Ensures that we save a weight checkpoint with every full checkpoint as well
-        if self.ckpt is not None:
-            if self.weights is None:
-                self.weights = WeightCheckpointConfig()
-            # If not interval is specified, use the same interval as the full checkpoint
-            if self.ckpt.interval is not None and self.weights.interval is None:
-                self.weights.interval = self.ckpt.interval
-            # If an interval is specified, ensure that the weight checkpoint interval is a multiple of the full checkpoint interval
-            if (
-                self.ckpt.interval is not None
-                and self.weights.interval is not None
-                and self.ckpt.interval % self.weights.interval != 0
-            ):
-                raise ValueError(
-                    "Use a weight checkpoint interval that ensures that a weight checkpoint is saved with every full checkpoint"
-                )
-        return self
-
-    @model_validator(mode="after")
     def validate_lora_adapter_saving(self):
-        if self.weights and self.weights.save_adapter_separately:
+        if self.ckpt and self.ckpt.weights and self.ckpt.weights.save_adapter_separately:
             lora_enabled = self.model and self.model.experimental and self.model.experimental.lora
             if not lora_enabled:
                 raise ValueError(
